requirements.txt
asgiref==3.8.1
Django==5.1.4
django-bootstrap5==24.3
psycopg2==2.9.10
sqlparse==0.5.3
typing_extensions==4.12.2
pillow==10.2.0
requests==2.32.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
